Module sodium_sys::crypto::aead

Secret Key Authenticated Encryption with Additional Data

This operation:

Encrypts a message with a key and a nonce to keep it confidential

Computes an authentication tag. This tag is used to make sure that the message, as well as optional, non-confidential (non-encrypted) data, haven't been tampered with. A typical use case for additional data is to store protocol-specific metadata about the message, such as its length and encoding.

The chosen construction uses encrypt-then-MAC and decryption will never be performed, even partially, before verification.

Constants

ABYTES
16 bytes for chacha20poly1305.

KEYBYTES
32 bytes for chacha20poly1305.

NPUBBYTES
8 bytes for chacha20poly1305.

PRIMITIVE
chacha20poly1305.

Functions

aead_decrypt
The aead_decrypt() function verifies that the ciphertext (as produced by aead_encrypt()) includes a valid tag using a secret key, a public nonce, and optional additional data.

aead_encrypt
The aead_encrypt() function encrypts a message using a secret key and a public nonce.