Function sodium_sys::crypto::aead::aead_decrypt

pub fn aead_decrypt<'a>(ciphertext: &[u8], add_data: Option<&[u8]>, nonce: &[u8], key: &[u8]) -> Result<&'a mut [u8], SSError>

The aead_decrypt() function verifies that the ciphertext (as produced by aead_encrypt()) includes a valid tag using a secret key, a public nonce, and optional additional data.

Additional data is optional and can be None if no additional data are required. Also, the additional data should be less than ABYTES in length.

The function returns a result containing decrypted message.

Examples