Module sodium_sys::crypto::asymmetrickey

Modules

auth_keypair
Memory safe keypair implementation.

authenc
Using public-key authenticated encryption, Bob can encrypt a confidential message specifically for Alice, using Alice's public key.

sealbox
Sealed boxes are designed to anonymously send messages to a recipient given its public key.

sign
In this system, a signer generates a key pair:

sign_keypair
Memory safe keypair implementation.