Module sodium_sys::crypto::asymmetrickey::authenc

Using public-key authenticated encryption, Bob can encrypt a confidential message specifically for Alice, using Alice's public key.

Using Bob's public key, Alice can verify that the encrypted message was actually created by Bob and was not tampered with, before eventually decrypting it.

Alice only needs Bob's public key, the nonce and the ciphertext. Bob should never ever share his secret key, even with Alice.

And in order to send messages to Alice, Bob only needs Alice's public key. Alice should never ever share her secret key either, even with Bob.

Alice can reply to Bob using the same system, without having to generate a distinct key pair.

The nonce doesn't have to be confidential, but it should be used with just one invokation of open_easy() for a particular pair of public and secret keys.

One easy way to generate a nonce is to use the Nonce struct, considering the size of nonces the risk of any random collisions is negligible. For some applications, if you wish to use nonces to detect missing messages or to ignore replayed messages, it is also ok to use a simple incrementing counter as a nonce.

When doing so you must ensure that the same value can never be re-used (for example you may have multiple threads or even hosts generating messages using the same key pairs).

This system provides mutual authentication. However, a typical use case is to secure communications between a server, whose public key is known in advance, and clients connecting anonymously.

Constants

BEFORENMBYTES

BOXZEROBYTES

MACBYTES

NONCEBYTES

PUBLICKEYBYTES

SECRETKEYBYTES

SEEDBYTES

ZEROBYTES

Functions

open
The open() function verifies and decrypts a ciphertext produced by seal().

open_detached
This function verifies and decrypts an encrypted message after verifying the given mac, and returns the decrypted message result.

open_detached_with_ssk
This function verifies and decrypts an encrypted message after verifying the given mac, and returns the decrypted message result.

open_nacl
The open_nacl() function decrypts a ciphertext produced by seal_nacl().

open_nacl_with_ssk
The open_nacl_with_ssk() function decrypts a ciphertext produced by seal_nacl_with_ssk().

open_with_ssk
The open_with_ssk() function verifies and decrypts a ciphertext produced by seal_with_ssk().

seal
The seal() function encrypts a message with a recipient's public key, a sender's secret key and a nonce.

seal_detached
This function encrypts a message with a recipients public key, your secret key and a nonce, and returns a tuple of byte arrays. The first element is the ciphertext, the second is the mac.

seal_detached_with_ssk
This function encrypts a message with a shared secret key, and a nonce, and returns a tuple of byte arrays. The first element is the ciphertext, the second is the mac.

seal_nacl
The seal_nacl() function encrypts a message with a recipient's public key, a sender's secret key and a nonce.

seal_nacl_with_ssk
The seal_nacl_with_ssk() function encrypts a message with a shared secret key and a nonce.

seal_with_ssk
The seal_with_ssk() function encrypts a message with a pre-calculated shared secret key and a nonce.

Type Definitions

SealDetachedResult