Module sodium_sys::crypto::asymmetrickey::sealbox

Sealed boxes are designed to anonymously send messages to a recipient given its public key.

Only the recipient can decrypt these messages, using its private key. While the recipient can verify the integrity of the message, it cannot verify the identity of the sender.

A message is encrypted using an ephemeral key pair, whose secret part is destroyed right after the encryption process.

Without knowing the secret key used for a given message, the sender cannot decrypt its own message later. And without additional data, a message cannot be correlated with the identity of its sender.

Constants

BOXZEROBYTES

MACBYTES

PUBLICKEYBYTES

SEALBYTES

SECRETKEYBYTES

ZEROBYTES

Functions

open
The open() function decrypts the ciphertext using the key pair (pk, sk), returns the message.

seal
The seal() function encrypts a message for a recipients public key. It returns the ciphertext.