Module sodium_sys::crypto::asymmetrickey::sign

In this system, a signer generates a key pair:

A secret key, that will be used to append a signature to any number of messages

A public key, that anybody can use to verify that the signature appended to a message was actually issued by the creator of the public key.

Verifiers need to already know and ultimately trust a public key before messages signed using it can be verified.

Warning: this is different from authenticated encryption. Appending a signature does not change the representation of the message itself.

Constants

BYTES

PUBLICKEYBYTES

SECRETKEYBYTES

SEEDBYTES

Functions

open
The open() function verifies a signature and returns the message with the signature stripped if successful.

open_detached
The open_detached() function verifies a signature is valid for a message and returns 0 on success, or -1 otherwise.

sign
The sign() function prepends a signature to a message, using a secret key.

sign_detached
The sign_detached() function generates a signature for message, using a secret key.