Function sodium_sys::crypto::asymmetrickey::sign::open_detached

pub fn open_detached(message: &[u8], signature: &[u8], pk: &[u8]) -> i32

The open_detached() function verifies a signature is valid for a message and returns 0 on success, or -1 otherwise.

pk is the public key of the sender that signed the message.

Examples

use sodium_sys::crypto::utils::init;
use sodium_sys::crypto::asymmetrickey::{sign,sign_keypair};

// Initialize sodium_sys
init::init();

// Create the keypair and activate for use.
let keypair = sign_keypair::KeyPair::new().unwrap();
keypair.activate_sk();
keypair.activate_pk();

// Generate the signature.
let signature = sign::sign_detached(b"test", keypair.sk_bytes()).unwrap();
let res = sign::open_detached(b"test", signature, keypair.pk_bytes());

assert!(res == 0);