Struct sodium_sys::crypto::asymmetrickey::sign_keypair::KeyPair

pub struct KeyPair {
    // some fields omitted
}

The key structure contains information necessary to create slices from raw parts.

Methods

impl KeyPair

fn new() -> Result<KeyPair, SSError>

Create a new keypair. The keys are generated with the crypto_sign_keypair() function to ensure safety and then set to no access via mprotect_noaccess() to ensure the data is not inadvertently (or maliciously) altered. Note in order to use the keypair, the caller must use activate_sk() and activate_pk().

Examples

use sodium_sys::crypto::utils::{init,secmem};
use sodium_sys::crypto::asymmetrickey::{sign,sign_keypair};

// Initialize the sodium-sys library.
init::init();

// Create a keypair for the box_ module.
let keypair = sign_keypair::KeyPair::new().unwrap();

// Activate the keys for use (they are created as no access).
keypair.activate_sk();
keypair.activate_pk();

// Validate.
assert!(keypair.sk_bytes().len() == sign::SECRETKEYBYTES);
assert!(secmem::memcmp(keypair.sk_bytes(),
                       &[0; sign::SECRETKEYBYTES]) != 0);
assert!(keypair.pk_bytes().len() == sign::PUBLICKEYBYTES);
assert!(keypair.pk_bytes() != [0; sign::PUBLICKEYBYTES]);

fn new_with_seed(seed: &[u8]) -> Result<KeyPair, SSError>

Create a new keypair with the given seed key. The keys are generated with the crypto_sign_seed_keypair() function to ensure safety and then set to no access via mprotect_noaccess() to ensure the data is not inadvertently (or maliciously) altered. Note in order to use the keypair, the caller must use activate_sk() and activate_pk().

Examples

use sodium_sys::crypto::utils::{init,secmem};
use sodium_sys::crypto::asymmetrickey::{sign,sign_keypair};

// Initialize the sodium-sys library.
init::init();

// Test seed key (don't use all zeros, it's a bad idea).
const TEST_SEED_KEY: [u8; sign::SEEDBYTES] = [0; sign::SEEDBYTES];

// Create a keypair for the box_ module.
let keypair = sign_keypair::KeyPair::new_with_seed(&TEST_SEED_KEY).unwrap();

// Activate the keys for use (they are created as no access).
keypair.activate_sk();
keypair.activate_pk();

// Validate.
assert!(keypair.sk_bytes().len() == sign::SECRETKEYBYTES);
assert!(secmem::memcmp(keypair.sk_bytes(),
                       &[0; sign::SECRETKEYBYTES]) != 0);
assert!(keypair.pk_bytes().len() == sign::PUBLICKEYBYTES);
assert!(keypair.pk_bytes() != [0; sign::PUBLICKEYBYTES]);

fn sk_bytes(&self) -> &[u8]

Convert the secret key to a byte sequence.

fn sk_bytes_mut(&self) -> &mut [u8]

Convert the secret key to a mutable byte sequence.

fn activate_sk(&self)

Activate the secret key for use via mprotect_readonly(). Note that once a secret key is created it cannot be modified in memory, only read.

fn deactivate_sk(&self)

De-activate the secret key via mprotect_noaccess(). Use this when the key isn't currently being used, but may be at a later time.

fn pk_bytes(&self) -> &[u8]

Convert the public key to a byte sequence.

fn pk_bytes_mut(&self) -> &mut [u8]

Convert the public key to a mutable byte sequence.

fn activate_pk(&self)

Activate the public key for use via mprotect_readonly(). Note that once a public key is created it cannot be modified in memory, only read.

fn deactivate_pk(&self)

De-activate the public key via mprotect_noaccess(). Use this when the key isn't currently being used, but may be at a later time.

fn get_seed<'a>(&self) -> Result<&'a [u8], SSError>

fn get_pk<'a>(&self) -> Result<&'a [u8], SSError>