Module sodium_sys::crypto::hash

Modules

generichash
This module computes a fixed-length fingerprint for an arbitrary long message.

passhash
Secret keys used to encrypt or sign confidential data have to be chosen from a very large keyspace. However, passwords are usually short, human-generated strings, making dictionary attacks practical.

sha2
The SHA-256 and SHA-512 functions are provided for interoperability with other applications.

shorthash
Many applications and programming language implementations were recently found to be vulnerable to denial-of-service attacks when a hash function with weak security guarantees, such as Murmurhash 3, was used to construct a hash table.