Module sodium_sys::crypto::hash::generichash

This module computes a fixed-length fingerprint for an arbitrary long message.

Sample use cases:

File integrity checking

Creating unique identifiers to index arbitrary long data

Constants

BYTES

BYTES_MAX

BYTES_MIN

KEYBYTES

KEYBYTES_MAX

KEYBYTES_MIN

Functions

finalize
Finalize and return the sha256 multi-part hash.

hash
The hash() function calculates a fingerprint of the message. The output size can be chosen by the application.

init
Initialize the sha256 multi-part hash.

state_size
The state_size() function should be used in conjunction with utils::malloc() to allocate the memory for the generic hash state at runtime.

update
Update the sha256 multi-part hash.