Module sodium_sys::crypto::hash::passhash

Secret keys used to encrypt or sign confidential data have to be chosen from a very large keyspace. However, passwords are usually short, human-generated strings, making dictionary attacks practical.

The pwhash operation derives a secret key of any size from a password and a salt.

The generated key has the size defined by the application, no matter what the password length is.

The same password hashed with same parameters will always produce the same key.

The same password hashed with different salts will produce different keys.

The function deriving a key from a password and a salt is CPU intensive and intentionally requires a fair amount of memory. Therefore, it mitigates brute-force attacks by requiring a significant effort to verify each password.

Common use cases:

Protecting an on-disk secret key with a password,

Password storage, or rather: storing what it takes to verify a password without having to store the actual password.

Notes

Do not forget to initialize the library with init(). The pwhash functions will still work without doing so, but possibly way slower.

Do not use constants (including OPSLIMIT_* and MEMLIMIT_*) in order to verify a password. Save the parameters along with the hash instead, and use these saved parameters for the verification.

Alternatively, use pwhash() and pwhash_verify(), that automatically take care of including and extracting the parameters.

By doing so, passwords can be rehashed using different parameters if required later on.

Cleartext passwords should not stay in memory longer than needed.

It is highly recommended to use mlock() to lock memory regions storing cleartext passwords, and to call munlock() right after pwhash() and pwhash_verify() return.

By design, a password whose length is 65 bytes or more is reduced to SHA-256(password). This can have security implications if the password is present in another password database using raw, unsalted SHA-256. Or when upgrading passwords previously hashed with unsalted SHA-256 to scrypt.

If this is a concern, passwords should be pre-hashed before being hashed using scrypt:

Guidelines for choosing scrypt parameters

Start by determining how much memory can be used the scrypt function. What will be the highest number of threads/processes evaluating the function simultaneously (ideally, no more than 1 per CPU core)? How much physical memory is guaranteed to be available?

memlimit should be a power of 2. Do not use anything less than 16 Mb, even for interactive use.

Then, a reasonable starting point for opslimit is memlimit / 32.

Measure how long the scrypt function needs in order to hash a password. If this it is way too long for your application, reduce memlimit and adjust opslimit using the above formula.

If the function is so fast that you can afford it to be more computationally intensive without any usability issues, increase opslimit.

For online use (e.g. login in on a website), a 1 second computation is likely to be the acceptable maximum.

For interactive use (e.g. a desktop application), a 5 second pause after having entered a password is acceptable if the password doesn't need to be entered more than once per session.

For non-interactive use and infrequent use (e.g. restoring an encrypted backup), an even slower computation can be an option.

But the best defense against brute-force password cracking remains using strong passwords. Libraries such as passwdqc can help enforce this.

Constants

MEMLIMIT_INTERACTIVE

MEMLIMIT_SENSITIVE

OPSLIMIT_INTERACTIVE

OPSLIMIT_SENSITIVE

SALTBYTES

STRBYTES

STRPREFIX

Functions

keygen
The keygen() function derives a key of the given length from a password and a salt.

pwhash
The pwhash() function creates an ASCII encoded string, which includes:

pwhash_verify
The pwhash_verify() function verifies that the given hash is a valid password verification string (as generated by pwhash()) for the given password.

scrypt
The traditional, low-level scrypt API is also available