Function sodium_sys::crypto::hash::passhash::keygen

pub fn keygen<'a>(password: &'a [u8], keylen: usize, salt: &[u8], opslimit: Option<usize>, memlimit: Option<usize>) -> Result<&'a [u8], SSError>

The keygen() function derives a key of the given length from a password and a salt.

The computed key is returned.

opslimit represents a maximum amount of computations to perform. Raising this number will make the function require more CPU cycles to compute a key.

memlimit is the maximum amount of RAM that the function will use, in bytes. It is highly recommended to allow the function to use at least 16 megabytes.

For interactive, online operations, OPSLIMIT_INTERACTIVE and MEMLIMIT_INTERACTIVE provide a safe base line for these two parameters. However, using higher values may improve security.

For highly sensitive data, OPSLIMIT_SENSITIVE and MEMLIMIT_SENSITIVE can be used as an alternative. But with these parameters, deriving a key takes about 2 seconds on a 2.8 Ghz Core i7 CPU and requires up to 1 gigabyte of dedicated RAM.

The salt should be unpredictable. randombytes::random_byte_array() is the easiest way to fill the salt.

Keep in mind that in order to produce the same key from the same password, the same salt, and the same values for opslimit and memlimit have to be used. Therefore, these parameters have to be stored for each user.

Examples

use sodium_sys::crypto::utils::{init,randombytes};
use sodium_sys::crypto::hash::passhash;

// Initialize sodium_sys
init::init();

// Generate the hash.
let mut salt = [0; passhash::SALTBYTES];
randombytes::random_byte_array(&mut salt);
let key = passhash::keygen(b"test", 64, &salt, None, None).unwrap();
assert!(key.len() == 64);