Module sodium_sys::crypto::hash::sha2

The SHA-256 and SHA-512 functions are provided for interoperability with other applications.

These functions are not keyed and are thus deterministic. In addition, they are vulnerable to length extension attacks.

A message can be hashed in a single pass, but a streaming API is also available to process a message as a sequence of multiple chunks.

If you are looking for a generic hash function and not specifically SHA-2, using generichash() (BLAKE2b) might be a better choice.

Constants

SHA256_BYTES

SHA512_BYTES

Functions

finalize256
Examples

finalize512
Examples

hash256
The hash256() function creates the SHA-256 hash for the given message.

hash512
The hash512() function creates the SHA-256 hash for the given message.

init256
Examples

init512
Examples

state_size_256
The state_size_256() function should be used in conjunction with utils::malloc() to allocate the memory for the hash state for SHA256 at runtime.

state_size_512
The state_size_512() function should be used in conjunction with utils::malloc() to allocate the memory for the hash state for SHA256 at runtime.

update256
Examples

update512
Examples