Module sodium_sys::crypto::hash::shorthash

Many applications and programming language implementations were recently found to be vulnerable to denial-of-service attacks when a hash function with weak security guarantees, such as Murmurhash 3, was used to construct a hash table.

In order to address this, Sodium provides the crypto_shorthash() function, which outputs short but unpredictable (without knowing the secret key) values suitable for picking a list in a hash table for a given key.

This function is optimized for short inputs.

The output of this function is only 64 bits. Therefore, it should not be considered collision-resistant.

Use cases:

Hash tables

Probabilistic data structures such as Bloom filters

Integrity checking in interactive protocols

Constants

BYTES

KEYBYTES

Functions

hash
Compute a fixed-size (BYTES bytes) fingerprint for the message using the key k.