Module sodium_sys::crypto::secretbox

Secret Key Authenticated Encryption

Encrypts a message with a key and a nonce to keep it confidential Computes an authentication tag. This tag is used to make sure that the message hasn't been tampered with before decrypting it. A single key is used both to encrypt/sign and verify/decrypt messages. For this reason, it is critical to keep the key confidential. The nonce doesn't have to be confidential, but it should never ever be reused with the same key.

Modules

crypto_secretbox_xsalsa20poly1305
xsalsa20poly1305 constants

Constants

BOXZEROBYTES
16 bytes for xsalsa20poly1305 primitive.

KEYBYTES
32 bytes for xsalsa20poly1305 primitive.

MACBYTES
16 bytes for xsalsa20poly1305 prmitive.

NONCEBYTES
24 bytes for xsalsa20poly1305 primitive.

PRIMITIVE
xsalsa20poly1305

ZEROBYTES
32 bytes for xsalsa20poly1305 primitive.

Functions

open
The open() function verifies and decrypts a ciphertext produced by seal().

open_detached
This function verifies and decrypts an encrypted message after verifying the given mac, and returns the decrypted message result.

open_nacl
The open_nacl() function verifies and decrypts a ciphertext using a secret key and a nonce. The open_nacl() function returns a Result containing a byte sequence representing the plaintext.

seal
The seal() function encrypts a message with a key and a nonce.

seal_detached
This function encrypts a message with a key and a nonce, and returns a tuple of byte arrays. The first element is the ciphertext, the second is the mac.

seal_nacl
The seal_nacl() function encrypts and authenticates a message using a secret key and a nonce. The seal_nacl() function returns Result containing a byte sequence containing the ciphertext.