Module sodium_sys::crypto::symmetrickey::aead

Secret Key Authenticated Encryption with Associated Data

This operation:

Encrypts a message with a key and a nonce to keep it confidential

Computes an authentication tag. This tag is used to make sure that the message, as well as optional, non-confidential (non-encrypted) data, haven't been tampered with. A typical use case for associated data is to store protocol-specific metadata about the message, such as its length and encoding.

The chosen construction uses encrypt-then-MAC and decryption will never be performed, even partially, before verification.

Notes

The nonce is 64 bits long. In order to prevent nonce reuse, if a key is being reused, it is recommended to increment the previous nonce instead of generating a random nonce for each message.

The API conforms to the proposed API for the CAESAR competition.

The construction conforms to the ChaCha20 and Poly1305 based Cipher Suites for TLS draft, and Sodium's implementation is fully interoperable with other current implementations.

The ChaCha20 and Poly1305 for IETF protocols RFC published later slightly changes the construction. That construction is also available by building sodium-sys with the latest feature as aead_encrypt_ietf() and aead_decrypt_ietf(). It features a longer nonce, which is IETF_NONCEBYTES bytes long.

A high-level aead_()* API is intentionally not defined until the CAESAR competition is over.

Constants

ABYTES
16 bytes.

KEYBYTES
32 bytes.

NPUBBYTES
0 bytes. 8 bytes.

Functions

aead_decrypt
The aead_decrypt() function verifies that the ciphertext (as produced by aead_encrypt()) includes a valid tag using a secret key, a public nonce, and optional associated data.

aead_encrypt
The aead_encrypt() function encrypts a message using a secret key and a public nonce.