Function sodium_sys::crypto::symmetrickey::aead::aead_encrypt

pub fn aead_encrypt<'a>(message: &[u8], add_data: Option<&[u8]>, key: &[u8], nonce: &[u8]) -> Result<&'a mut [u8], SSError>

The aead_encrypt() function encrypts a message using a secret key and a public nonce.

The encrypted message, as well as a tag authenticating both the confidential message and non-confidential data, are put into the output byte sequence.

Associated data is optional and can be None if no associated data are required.. Also, the associated data should be less than ABYTES in length.

The public nonce should never ever be reused with the same key. The recommended way to generate it is to use the Nonce struct for the first message, and increment it for each subsequent message using the same key.

Examples

use sodium_sys::crypto::utils::{init,nonce};
use sodium_sys::crypto::symmetrickey::{aead,key};

// Initialize sodium_sys
init::init();

// Create the key and activate for use.
let key = key::Key::new(aead::KEYBYTES);
key.activate();

// Create the nonce and activate for use.
let nonce = nonce::Nonce::new(aead::NPUBBYTES);
nonce.activate();

// Generate the ciphertext and protect it as readonly.
let ciphertext = aead::aead_encrypt(b"test",
                                    None,
                                    key.bytes(),
                                    nonce.bytes()).unwrap();
println!("{:?}", ciphertext);

// Generate the ciphertext with additional data and protect it as readonly.
let ciphertext = aead::aead_encrypt(b"test",
                                    Some(b"more data"),
                                    key.bytes(),
                                    nonce.bytes()).unwrap();
println!("{:?}", ciphertext);