Module sodium_sys::crypto::symmetrickey::auth

Secret Key Message Authentication

This operation computes an authentication tag for a message and a secret key, and provides a way to verify that a given tag is valid for a given message and a key.

The function computing the tag deterministic: the same (message, key) tuple will always produce the same output.

However, even if the message is public, knowing the key is required in order to be able to compute a valid tag. Therefore, the key should remain confidential. The tag, however, can be public.

A typical use case is:

A prepares a message, add an authentication tag, sends it to B

A doesn't store the message

Later on, B sends the message and the authentication tag to A

A uses the authentication tag to verify that it created this message. This operation does not encrypt the message. It only computes and verifies an authentication tag.

Constants

BYTES
32 bytes.

KEYBYTES
32 bytes.

Functions

auth
The auth() function computes a tag for the message and a key. The key should be KEYBYTES bytes. The function return the tag byte sequence.

auth_verify
The auth_verify() function verifies that the mac is a valid mac for the given message and the key k.