Module sodium_sys::crypto::symmetrickey::authenc

Authenticated Encryption Symmetric (Secret) Key Authenticated Encryption

Encrypts a message with a key and a nonce to keep it confidential Computes an authentication tag. This tag is used to make sure that the message hasn't been tampered with before decrypting it. A single key is used both to encrypt/sign and verify/decrypt messages. For this reason, it is critical to keep the key confidential. The nonce doesn't have to be confidential, but it should never ever be reused with the same key.

Constants

BOXZEROBYTES
16 bytes.

KEYBYTES
32 bytes.

MACBYTES
ZEROBYTES - BOXZEROBYTES (16 bytes).

NONCEBYTES
24 bytes.

ZEROBYTES
32 bytes.

Functions

encrypt
The encrypt() function encrypts a message with a key and a nonce.

encrypt_detached
This function encrypts a message with a key and a nonce, and returns a tuple of byte arrays. The first element is the ciphertext, the second is the mac.

encrypt_nacl
The encrypt_nacl() function encrypts and authenticates a message using a secret key and a nonce. The encrypt_nacl() function returns Result containing a byte sequence containing the ciphertext.

open
The open() function verifies and decrypts a ciphertext produced by encrypt().

open_detached
This function verifies and decrypts an encrypted message after verifying the given mac, and returns the decrypted message result.

open_nacl
The open_nacl() function verifies and decrypts a ciphertext using a secret key and a nonce. The open_nacl() function returns a Result containing a byte sequence representing the plaintext.

Type Definitions

EncryptDetachedResult