Struct sodium_sys::crypto::symmetrickey::key::Key

pub struct Key {
    // some fields omitted
}

The key structure contains information necessary to create slices from raw parts.

Methods

impl Key

fn new(size: usize) -> Key

Create a new key of the given size. The key is malloc'd with the libsodium malloc function to ensure safety, filled with random bytes via random_byte_array(), and then set to no access via mprotect_noaccess() to ensure the data is not inadvertently (or maliciously) altered. Note in order to use the key, the caller must use activate().

fn bytes(&self) -> &[u8]

Convert the key to a byte sequence.

fn bytes_mut(&self) -> &mut [u8]

Convert the key to a mutable byte sequence.

fn activate(&self)

Activate the key for use via mprotect_readonly(). Note that once a key is created it cannot be modified in memory, only read.

fn deactivate(&self)

De-activate the key via mprotect_noaccess(). Use this when the key isn't currently being used, but may be at a later time.