Function sodium_sys::crypto::utils::secmem::free

pub fn free(mem: &[u8])

The free() function unlocks and deallocates memory allocated using malloc() or allocarray().

Prior to this, the canary is checked in order to detect possible buffer underflows and terminate the process if required.

free() also fills the memory region with zeros before the deallocation.

This function can be called even if the region was previously protected using mprotect_readonly(); the protection will automatically be changed as needed.

The free() function safely wraps the sodium_free function.

Examples

use sodium_sys::crypto::utils::init::init;
use sodium_sys::crypto::utils::secmem::{malloc,free};

let _ = init();
let mut v = malloc(128);
v[0] = 1;
v[127] = 255;
assert!(v.len() == 128);
assert!(v[0] == 1);
assert!(v[127] == 255);
free(v);