Function sodium_sys::crypto::utils::secmem::hex2bin

pub fn hex2bin(hex: String, output: &mut Vec<u8>, ignore: Option<String>) -> i32

The hex2bin() function parses a hexadecimal string and converts it to a byte sequence.

ignore is a string of characters to skip. For example, the string ": " allows columns and spaces to be present at any locations in the hexadecimal string. These characters will just be ignored. As a result, "69:FC", "69 FC", "69 : FC" and "69FC" will be valid inputs, and will produce the same output.

ignore can be set to None in order to disallow any non-hexadecimal character.

The function returns -1 on failure. It returns 0 on success and sets output to the byte sequence.

It evaluates in constant time for a given length and format.

hex2bin() safely wraps the sodium_hex2bin() function.

Examples

use sodium_sys::crypto::utils::secmem::hex2bin;

let hex = String::from("0001feff");
let mut output = Vec::new();
assert!(hex2bin(hex, &mut output, None) == 0);
assert!(output == [0, 1, 254, 255]);

let hex = String::from("00:01:fe:ff");
let ignore = Some(String::from(":"));
let mut output = Vec::new();
assert!(hex2bin(hex, &mut output, ignore) == 0);
assert!(output == [0, 1, 254, 255]);

let hex = String::from("00 01 fe ff");
let ignore = Some(String::from(" "));
let mut output = Vec::new();
assert!(hex2bin(hex, &mut output, ignore) == 0);
assert!(output == [0, 1, 254, 255]);

let hex = String::from("00 01:fe ff");
let ignore = Some(String::from(": "));
let mut output = Vec::new();
assert!(hex2bin(hex, &mut output, ignore) == 0);
assert!(output == [0, 1, 254, 255]);