Function sodium_sys::crypto::utils::secmem::malloc

pub fn malloc<'a>(size: usize) -> &'a mut [u8]

The malloc() function returns a mutable array of bytes.

The allocated region is placed at the end of a page boundary, immediately followed by a guard page. As a result, accessing memory past the end of the region will immediately terminate the application.

A canary is also placed right before the returned pointer. Modification of this canary are detected when trying to free the allocated region with free(), and also cause the application to immediately terminate.

An additional guard page is placed before this canary to make it less likely for sensitive data to be accessible when reading past the end of an unrelated region.

The allocated region is filled with 0xd0 bytes in order to help catch bugs due to initialized data.

In addition, sodium_mlock() is called on the region to help avoid it being swapped to disk. On operating systems supporting MAP_NOCORE or MADV_DONTDUMP, memory allocated this way will also not be part of core dumps.

The returned address will not be aligned if the allocation size is not a multiple of the required alignment.

For this reason, malloc() should not be used with packed or variable-length structures, unless the size given to malloc() is rounded up in order to ensure proper alignment.

All the structures used by libsodium can safely be allocated using sodium_malloc(), the only one requiring extra care being crypto_generichash_state, whose size needs to be rounded up to a multiple of 64 bytes.

Examples

use sodium_sys::crypto::utils::init::init;
use sodium_sys::crypto::utils::secmem::{malloc,free};

let _ = init();
let mut v = malloc(64);
v[0] = 1;
assert!(v.len() == 64);
assert!(v[0] == 1);
free(v);