Function sodium_sys::crypto::utils::secmem::mprotect_noaccess

pub fn mprotect_noaccess(mem: &[u8])

The mprotect_noaccess() function makes a region allocated using malloc() or allocarray() inaccessible. It cannot be read or written, but the data are preserved.

This function can be used to make confidential data inacessible except when actually needed for a specific operation.

mprotect_noaccess() safely wraps the sodium_mprotect_noaccess() function.

Examples

use sodium_sys::crypto::utils::init::init;
use sodium_sys::crypto::utils::secmem::{free,malloc,mprotect_noaccess};

let _ = init();
let mut v = malloc(64);
v[0] = 1;
assert!(v.len() == 64);
assert!(v[0] == 1);
mprotect_noaccess(&mut v);
// If you uncomment the following line the program will fail (no read).
// assert!(v[0] == 1);
// If you uncomment the following line the program will fail (no write).
// v[1] = 1;
free(&mut v);