Function sodium_sys::utils::mprotect_readwrite

pub fn mprotect_readwrite(mem: &[u8])

The mprotect_readwrite() function marks a region allocated using malloc() or allocarray() as readable and writable, after having been protected using mprotect_readonly() or mprotect_noaccess().

Examples

use sodium_sys::core::init;
use sodium_sys::utils::{free,malloc,mprotect_noaccess,mprotect_readwrite};

let _ = init();
let mut v = malloc(64);
v[0] = 1;
assert!(v.len() == 64);
assert!(v[0] == 1);
mprotect_noaccess(&mut v);
// If you uncomment the following line the program will fail (no read).
// assert!(v[0] == 1);
// If you uncomment the following line the program will fail (no write).
// v[1] = 1;
mprotect_readwrite(&mut v);
assert!(v[0] == 1);
v[1] = 1;
assert!(v[1] == 1);
free(&mut v);